backend/requirements.txt
opencv-python
fastapi
uvicorn
pillow
torch
torchvision
efficientnet_pytorch
scikit-learn
numpy
scikit-learn
